backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.15.0
openai==0.27.0
python-dotenv==0.19.0
setuptools>=58.0.0
flask-cors==3.0.10